Calcul des métadonnées GBFS lors de la validation d'un dataset  (#1992)
* WIP

* Remove defmock

* Move to shared
