+++ apps/db/test/resource_test.exs
@@ -41,6 +41,16 @@ defmodule DB.ResourceTest do
     assert log_validation.skipped_reason == "no previous validation"
   end
 
+  test "validate and save a GBFS resource" do
+    resource = insert(:resource, %{url: "url1", format: "gbfs"})
+
+    Transport.Shared.GBFSMetadata.Mock
+    |> expect(:compute_feed_metadata, fn _resource, _cors_base_url -> %{"foo" => "bar"} end)
+
+    assert Resource.validate_and_save(resource, false) == {:ok, nil}
+    assert %{metadata: %{"foo" => "bar"}} = Repo.get(Resource, resource.id)
+  end
+
   test "validation is skipped if previous validation is still valid" do
     resource = insert(:resource, %{url: "url1", format: "GTFS", content_hash: "sha256_hash"})
