+++ apps/transport/lib/transport_web/controllers/gbfs_analyzer_controller.ex
@@ -2,7 +2,7 @@ defmodule TransportWeb.GbfsAnalyzerController do
   use TransportWeb, :controller
 
   def index(conn, %{"url" => gbfs_url}) when gbfs_url !== "" do
-    metadata = Transport.GBFSMetadata.Wrapper.compute_feed_metadata(gbfs_url)
+    metadata = Transport.GBFSMetadata.compute_feed_metadata(gbfs_url)
 
     conn
     |> assign(:gbfs_url, gbfs_url)
