+++ apps/transport/test/support/mocks.ex
@@ -4,5 +4,5 @@ Mox.defmock(Validation.Validator.Mock, for: Shared.Validation.Validator)
 Mox.defmock(Shared.Validation.GBFSValidator.Mock, for: Shared.Validation.GBFSValidator.Wrapper)
 Mox.defmock(Transport.Rambo.Mock, for: Transport.RamboLauncher)
 Mox.defmock(Transport.Notifications.FetcherMock, for: Transport.Notifications.Fetcher)
-Mox.defmock(Transport.GBFSMetadata.Mock, for: Transport.GBFSMetadata.Wrapper)
+Mox.defmock(Transport.Shared.GBFSMetadata.Mock, for: Transport.Shared.GBFSMetadata.Wrapper)
 Mox.defmock(Transport.AvailabilityChecker.Mock, for: Transport.AvailabilityChecker.Wrapper)
