+++ apps/transport/test/transport_web/controllers/gbfs_analyzer_controller_test.exs
@@ -19,8 +19,10 @@ defmodule TransportWeb.GbfsAnalyzerControllerTest do
     test "with a GBFS to analyze", %{conn: conn} do
       gbfs_url = "/gbfs.json"
 
-      Transport.GBFSMetadata.Mock
-      |> expect(:compute_feed_metadata, fn ^gbfs_url ->
+      Transport.Shared.GBFSMetadata.Mock
+      |> expect(:compute_feed_metadata, fn ^gbfs_url, cors_base_url ->
+        assert cors_base_url == TransportWeb.Endpoint.url()
+
         %{
           system_details: %{name: "GBFS feed name"},
           feeds: ["station_information"]
