+++ config/config.exs
@@ -108,7 +108,7 @@ config :transport,
   history_impl: Transport.History.Fetcher.S3,
   gtfs_to_json_converter_url: "https://convertisseur.transport.data.gouv.fr/gtfs2geojson_sync",
   rambo_impl: Transport.Rambo,
-  gbfs_metadata_impl: Transport.GBFSMetadata,
+  gbfs_metadata_impl: Transport.Shared.GBFSMetadata,
   availability_checker_impl: Transport.AvailabilityChecker
 
 config :datagouvfr,
@@ -128,7 +128,10 @@ config :ex_aws,
   ],
   json_codec: Jason
 
+domain_name = System.get_env("DOMAIN_NAME", "transport.data.gouv.fr")
+
 config :transport,
+  domain_name: domain_name,
   max_import_concurrent_jobs: (System.get_env("MAX_IMPORT_CONCURRENT_JOBS") || "1") |> String.to_integer(),
   nb_days_to_keep_validations: 60,
   join_our_slack_link: "https://join.slack.com/t/transportdatagouvfr/shared_invite/zt-2n1n92ye-sdGQ9SeMh5BkgseaIzV8kA",
